# Perf(observableset): Remove extra clone operation on .values .keys .Iterator access