--- a/.changeset/odd-dogs-thank.md
+++ b/.changeset/odd-dogs-thank.md
@@ -1,0 +1,5 @@
+---
+"mobx": patch
+---
+
+Improve observableset memory footprint and performance