<!--
    Thanks for taking the effort to create a PR! 🙌

    👋 Are you making a change to documentation only? Delete the rest of the template and go ahead.

    👋 If you are creating an extensive PR, you might want to open an issue with your idea first in case there is a chance for rejecting it.

    👋 If you intend to work on PR over several days, please, create [draft pull requests](https://github.blog/2019-02-14-introducing-draft-pull-requests/) instead.

    👇 Please look at the following checklist to ensure that your PR can be accepted quickly:
-->

### Code change checklist

-   [ ] Added/updated unit tests
-   [ ] Updated `/docs`. For new functionality, at least `API.md` should be updated
-   [x] Verified that there is no significant performance drop (`yarn mobx test:performance`)

<!--
    Feel free to ask help with any of these boxes!
-->

### Context
within the observableset, on retriving `entries/keys/values` or during iteration, there is an extra `Array.from` operation that is causing an unwanted copy of values within memory, the significance of extra operation is amplified when considering iterator helper methods where it would read the entire range of elements within the set instead of lazy evaluation, i.e: `Set.prototype.entries().take(1)`, or any other case when the iterator is partially retrieved `Set.prototype.entries().next()`.

### Benchmarks
#### Time
Use a macbook air M1 - 8GB model - plugged in, 1000 iterations
| test |  before |  after | note |
|---|---|---|---|
|  Set: conversion to array - 1000 elements | 3.33s min - 3.41s max | 2.90s min - 2.95s max  | non significant |
|  Set: conversion to array - 10000 elements | 12.74s min - 12.87s max | 9.10s min - 9.16s max | non significant |
|  Set: iterator helpers - 1000 elements, take(1) | 0.60s min - 0.77s max | 0.31s min - 0.32s max | exponential 🎊  |
|  Set: iterator helpers - 10000 elements, take(1) | 4.19s min - 4.22s max | 0.31s min - 0.32s max | exponential 🎊   |

#### Memory
Mesuring the effect on memory when converting a large observableset to an array.

| size |  before |  after |
|---|---|---|
|  1M item |  66 Mb | 46 Mb | 
|  10M item | 617 Mb | 423 Mb | 

```
   function benchmarkMemoryUsage() {
            const beforeMemory = memoryUsage().heapUsed
            
            const obs = observable.set()
            for (let i = 0; i < 10000_000; i++) {
                obs.add(i)
            }

            const x = Array.from(obs)
            const afterMemory1 = memoryUsage().heapUsed

            console.log(`Memory diff: ${(afterMemory1 - beforeMemory) / 1000_000} Mbytes %`)
        }
```