@dkazakov8 2025-02-01
Nice work! 

Is there a specific reason to use `this.values()`

```
entries() {
  const values = this.values()
	
  return makeIterableForSet<[T, T]>({
    next() {
      const { value, done } = values.next()
	  
      return !done ? { value: [value, value], done } : { value: undefined, done }
    }
  })
}
```

instead of `this.entries()`?

```
entries() {
  const values = this.entries()
	
  return makeIterableForSet<[T, T]>({
    next() { return values.next() }
  })
}
```
	
Seems like the second approach would be more optimized (less constants, less objects, destructuring and no manual creation of arrays)